Revert "ACP-27 Add snapshot tests (#17)"
This reverts commit a955a623eb62a1dece2da5f12f1a7e5721f3c2bc.
