--- .gitattributes
@@ -1,1 +1,0 @@
-**/snapshots/**/*.png filter=lfs diff=lfs merge=lfs -text
