+++ build-logic/convention/src/main/java/com/ignacnic/convention/ComposeConventionPlugin.kt
@@ -39,7 +39,6 @@ class ComposeConventionPlugin : Plugin<Project> {
             add("implementation", libs.findLibrary("androidx-material3").get())
             add("implementation", libs.findLibrary("androidx-ui-test-manifest").get())
             add("implementation", libs.findLibrary("androidx-ui-test-junit4").get())
-            add("debugImplementation", libs.findLibrary("androidx-ui-tooling").get())
         }
     }
 }
