+++ business-logic/elevation/build.gradle.kts
@@ -1,7 +1,7 @@
 plugins {
     alias(libs.plugins.ignacnic.android.library)
     alias(libs.plugins.kotlinx.serialization)
-    alias(libs.plugins.ignacnic.test.unit)
+    alias(libs.plugins.ignacnic.test)
 }
 
 android {
