+++ business-logic/gpx-file/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     alias(libs.plugins.ignacnic.kotlin.library)
-    alias(libs.plugins.ignacnic.test.unit)
+    alias(libs.plugins.ignacnic.test)
 }
 
 dependencies {
