+++ common/ui-resources/build.gradle.kts
@@ -1,8 +1,6 @@
 plugins {
     alias(libs.plugins.ignacnic.android.library)
     alias(libs.plugins.ignacnic.compose)
-    alias(libs.plugins.ignacnic.test.screenshot)
-    alias(libs.plugins.paparazzi)
 }
 
 android {
