--- common/ui-resources/src/test/snapshots/images/com.ignacnic.architectcoders.common.uiresources.components_RationaleDialogTest_testRationaleDialog.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a87c7dfcef7ebe4d27dc6f9d5a311562e484f24bbd175b55c27c9553e7b2f5d2
-size 27313
