+++ feature/location/build.gradle.kts
@@ -1,9 +1,7 @@
 plugins {
     alias(libs.plugins.ignacnic.android.library)
     alias(libs.plugins.ignacnic.compose)
-    alias(libs.plugins.ignacnic.test.unit)
-    alias(libs.plugins.ignacnic.test.screenshot)
-    alias(libs.plugins.paparazzi)
+    alias(libs.plugins.ignacnic.test)
 }
 
 android {
