--- feature/location/src/test/snapshots/images/com.ignacnic.architectcoders.feature.location.detail_LocationDetailScreenTest_testErrorDetailScreen.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7d1e64c8155b72b7826952e92e2ea6946d3ca81768f24fd0ab44ff5f83c0e3fe
-size 33061
