--- feature/location/src/test/snapshots/images/com.ignacnic.architectcoders.feature.location.detail_LocationDetailScreenTest_testLoadedDetailScreen.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:02830197a82adeb56cf6850851f57116d30a2ae1bd5b9d3588d70fd67227dc32
-size 27790
