--- feature/location/src/test/snapshots/images/com.ignacnic.architectcoders.feature.location.detail_LocationDetailScreenTest_testLoadingDetailScreen.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:cd24469232e3557a5ee24889a627e9f8c8fa89261569c992b4f9a6128462b681
-size 25975
