--- feature/location/src/test/snapshots/images/com.ignacnic.architectcoders.feature.location.requester_LocationRequesterScreenTest_testEmptyRequesterScreen.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2121f4592d9ab11510cf3d9f4d4bbe3776ef74624f23ccebc1c2fba4b3c985a5
-size 11466
