--- feature/location/src/test/snapshots/images/com.ignacnic.architectcoders.feature.location.requester_LocationRequesterScreenTest_testFilledRequesterScreen.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f7acd7079462e886ed77424e82ed72748f9857c0ba1340b71a48646dfacb47fc
-size 23529
